Apparently scout vehicles have been broken for years, so that's fixed now
Took so long to debug. Previously once the scout vehicles do 1 loop around the target island, they would have no more points so they would get put into holding state. Then they would be forever stuck in the holding state even after finishing the scouting and the code in the "vehicle_object.role == "scout"" branch never gets used.

In the history and when the scout code was originally written, the if statement was originally #vehicle_object.path > 1 but it later got changed to >0 and then doing nextPath and another check for >0. Its always greater than 0 before nextPath since it would error so the fix was to remove that first if statement and move everything down after nextPath gets used so it properly runs if its reached the last point in the path and its a scout vehicle.